fixed 2.1 and added 2.2

+++ assignments/assignment2/gradient_check.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 
-def check_gradient(f, x, delta=1e-5, tol=1e-4):
+def check_gradient(f, x, delta=1e-5, tol = 1e-4):
     '''
     Checks the implementation of analytical gradient by comparing
     it to numerical gradient using two-point formula
@@ -15,7 +15,7 @@ def check_gradient(f, x, delta=1e-5, tol=1e-4):
     '''
     
     assert isinstance(x, np.ndarray)
-    assert x.dtype == float
+    assert x.dtype == np.float
     
     orig_x = x.copy()
     fx, analytic_grad = f(x)
@@ -35,6 +35,7 @@ def check_gradient(f, x, delta=1e-5, tol=1e-4):
         xdelta[ix] = delta
         numeric_grad_at_ix = (f(x + xdelta)[0] - f(x - xdelta)[0]) / (2 * delta)
         
+        # TODO compute value of numeric gradient of f to idx
         if not np.isclose(numeric_grad_at_ix, analytic_grad_at_ix, tol):
             print("Gradients are different at %s. Analytic: %2.5f, Numeric: %2.5f" % (ix, analytic_grad_at_ix, numeric_grad_at_ix))
             return False
@@ -48,13 +49,11 @@ def check_gradient(f, x, delta=1e-5, tol=1e-4):
 def check_layer_gradient(layer, x, delta=1e-5, tol=1e-4):
     """
     Checks gradient correctness for the input and output of a layer
-
     Arguments:
       layer: neural network layer, with forward and backward functions
       x: starting point for layer input
       delta: step to compute numerical gradient
       tol: tolerance for comparing numerical and analytical gradient
-
     Returns:
       bool indicating whether gradients match or not
     """
@@ -76,14 +75,12 @@ def check_layer_param_gradient(layer, x,
                                delta=1e-5, tol=1e-4):
     """
     Checks gradient correctness for the parameter of the layer
-
     Arguments:
       layer: neural network layer, with forward and backward functions
       x: starting point for layer input
       param_name: name of the parameter
       delta: step to compute numerical gradient
       tol: tolerance for comparing numerical and analytical gradient
-
     Returns:
       bool indicating whether gradients match or not
     """
@@ -109,14 +106,12 @@ def check_model_gradient(model, X, y,
                          delta=1e-5, tol=1e-4):
     """
     Checks gradient correctness for all model parameters
-
     Arguments:
       model: neural network model with compute_loss_and_gradients
       X: batch of input data
       y: batch of labels
       delta: step to compute numerical gradient
       tol: tolerance for comparing numerical and analytical gradient
-
     Returns:
       bool indicating whether gradients match or not
     """
